+++ sumo/src/netload/NLSucceedingLaneBuilder.h
@@ -99,7 +99,7 @@ public:
 #endif
                      MSLink::LinkDirection dir, MSLink::LinkState state,
                      bool internalEnd,
-                     const std::string &tlid="", size_t linkNo=0) throw(InvalidArgument);
+                     const std::string &tlid="", unsigned int linkNo=0) throw(InvalidArgument);
 
 
     /** @brief Ends the computation of a container holding the succeeding lanes of a lane
